шоп/requirements.txt
aiogram==3.13.1
aiohttp==3.10.10
aiosqlite==0.20.0
SQLAlchemy==2.0.36
Telethon==1.37.0
python-dotenv==1.0.1
